Adding exceptions, immuting classes, cleaning up imports and class signatures in model.

+++ src/Model/Service/LoanService.java
@@ -1,8 +1,6 @@
 package Model.Service;
 
-import Exceptions.DeleteFailedException;
-import Exceptions.FailedReturnException;
-import Exceptions.InsertFailedException;
+import Exceptions.*;
 import Model.Dao.BookDaoImpl;
 import Model.Dao.LoanDao;
 import Model.Dao.LoanDaoImpl;
@@ -11,7 +9,6 @@ import Model.Entity.Book;
 import Model.Entity.Loan;
 import Model.Entity.User;
 
-import javax.crypto.DecapsulateException;
 import java.time.LocalDate;
 import java.util.ArrayList;
 
@@ -28,17 +25,28 @@ public class LoanService {
         this.userDao = UserDaoImpl.getInstance();
     }
 
-    public Boolean insert(Loan loan) throws InsertFailedException, FailedReturnException {;
-        return loanDao.insert(loan) && bookDao.update(loan.getBook());
+    public Boolean insert(Loan loan) throws InsertFailedException, InvalidValuesException {
+        if (!loan.getBook().isAvailable() || loan.getBailee().getPenalty() > 0 || loan.getReturnDate().isBefore(LocalDate.now()))
+            throw new InvalidValuesException("Dados do emprestimo inválidos.");
+
+        return loanDao.insert(loan);
     }
-    public Boolean delete(Loan loan) throws DeleteFailedException {
+
+    public Boolean delete(Loan loan) throws DeleteFailedException, InvalidValuesException {
+        if (loan == null)
+            throw new InvalidValuesException("Emprestimo inválido.");
+
         return loanDao.delete(loan);
     }
+
     public List<Loan> getAll(){
         return loanDao.getAll();
     }
 
-    public List<Loan> getByBook(Book book) {
+    public List<Loan> getByBook(Book book) throws InvalidValuesException {
+        if (book == null)
+            throw new InvalidValuesException("Livro inexistente.");
+
         ArrayList <Loan> loans = new ArrayList<>();
 
         for (Loan l : loanDao.getAll()){
@@ -49,7 +57,10 @@ public class LoanService {
         return loans;
     }
 
-    public List<User> getByUser(User user) {
+    public List<User> getByUser(User user) throws InvalidValuesException {
+        if (user == null)
+            throw new InvalidValuesException("Usuário inexistente.");
+
         ArrayList <User> users = new ArrayList<>();
         for (User u : userDao.getAll()){
             if (u.getEmail().equals(user.getEmail())){
@@ -59,7 +70,10 @@ public class LoanService {
         return users;
     }
 
-    public List<Loan> getByReturnDate(LocalDate date) {
+    public List<Loan> getByReturnDate(LocalDate date) throws InvalidDateException{
+        if (date == null)
+            throw new InvalidDateException("Data de devolução invalida.");
+
         ArrayList <Loan> loans = new ArrayList<>();
         for (Loan l : loanDao.getAll()){
             if (l.getReturnDate().isEqual(date)){
@@ -69,7 +83,10 @@ public class LoanService {
         return loans;
     }
 
-    public List<Loan> getByAcquiredDate(LocalDate date) {
+    public List<Loan> getByAcquiredDate(LocalDate date) throws InvalidDateException {
+        if (date == null)
+            throw new InvalidDateException("Data de aquisição invalida.");
+
         ArrayList <Loan> loans = new ArrayList<>();
         for (Loan l : loanDao.getAll()){
             if (l.getAcquiredDate().isEqual(date)){
